# core/api/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateApiKeyRequest": {
        "properties": {
          "challenge_id": {
            "description": "The challenge id associated with the previously displayed code",
            "example": "67647f5ecda913e9a2e11b26",
            "type": "string"
          },
          "code": {
            "description": "The 4-digit code retrieved from Anytype Desktop app",
            "example": "1234",
            "type": "string"
          }
        },
        "type": "object"
      },
      "CreateApiKeyResponse": {
        "properties": {
          "api_key": {
            "description": "The api key used to authenticate requests",
            "example": "zhSG/zQRmgADyilWPtgdnfo1qD60oK02/SVgi1GaFt6=",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ServerError": {
        "properties": {
          "code": {
            "example": "internal_server_error",
            "type": "string"
          },
          "message": {
            "example": "Internal server error",
            "type": "string"
          },
          "object": {
            "example": "error",
            "type": "string"
          },
          "status": {
            "example": 500,
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "code": {
            "example": "bad_request",
            "type": "string"
          },
          "message": {
            "example": "Bad request",
            "type": "string"
          },
          "object": {
            "example": "error",
            "type": "string"
          },
          "status": {
            "example": 400,
            "type": "integer"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerauth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@anytype.io",
      "name": "Anytype Support",
      "url": "https://anytype.io/contact"
    },
    "description": "This API enables seamless interaction with Anytype's resources - spaces, objects, properties, types, templates, and beyond.",
    "license": {
      "name": "Any Source Available License 1.0",
      "url": "https://github.com/anyproto/anytype-api/blob/main/LICENSE.md"
    },
    "termsOfService": "https://anytype.io/terms_of_use",
    "title": "Anytype API",
    "version": "2025-11-08"
  },
...
  "paths": {
    "/v1/auth/api_keys": {
      "post": {
        "description": "After receiving a `challenge_id` from the `/v1/auth/challenges` endpoint, the client calls this endpoint to provide the corresponding 4-digit code along with the challenge ID. The endpoint verifies that the challenge solution is correct and, if it is, returns an `api_key`. This endpoint is central to the authentication process, as it validates the user's identity and issues a key that can be used for further interactions with the API.",
        "operationId": "create_api_key",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "description": "The request body containing the challenge ID and code",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiKeyResponse"
                }
              }
            },
            "description": "The API key that can be used in the Authorization header for subsequent requests"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Bad request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Create API Key",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/challenges": {
      "post": {
        "description": "Generates a one-time authentication challenge for granting API access to the user's vault. Upon providing a valid `app_name`, the server issues a unique `challenge_id` and displays a 4-digit code within the Anytype Desktop. The `challenge_id` must then be used with the `/v1/auth/api_keys` endpoint to solve the challenge and retrieve an authentication token. This mechanism ensures that only trusted applications and authorized users gain access.",
        "operationId": "create_auth_challenge",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChallengeRequest"
              }
            }
          },
          "description": "The request body containing the app name",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateChallengeResponse"
                }
              }
            },
            "description": "The challenge ID associated with the started challenge"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Bad request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Create Challenge",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/search": {
      "post": {
        "description": "Executes a global search over all spaces accessible to the authenticated user. The request body must specify the `query` text (currently matching only name and snippet of an object), optional filters on types (e.g., \"page\", \"task\"), and sort directives (default: descending by last modified date). Pagination is controlled via `offset` and `limit` query parameters to facilitate lazy loading in client UIs. The response returns a unified list of matched objects with their metadata and properties.",
        "operationId": "search_global",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          },
          {
            "description": "The number of items to skip before starting to collect the result set",
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "type": "integer"
            }
          },
          {
            "description": "The number of items to return",
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "maximum": 1000,
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "description": "The search parameters used to filter and sort the results",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse-Object"
                }
              }
            },
            "description": "The list of objects matching the search criteria"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "Search objects across all spaces",
        "tags": [
          "Search"
        ]
      }
    },
    "/v1/spaces": {
      "get": {
        "description": "Retrieves a paginated list of all spaces that are accessible by the authenticated user. Each space record contains detailed information such as the space ID, name, icon (derived either from an emoji or image URL), and additional metadata. This endpoint is key to displaying a user's workspaces.\nSupports dynamic filtering via query parameters (e.g., ?name[contains]=project). See FilterCondition enum for available conditions.",
        "operationId": "list_spaces",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          },
          {
            "description": "The number of items to skip before starting to collect the result set",
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "type": "integer"
            }
          },
          {
            "description": "The number of items to return",
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "maximum": 1000,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse-Space"
                }
              }
            },
            "description": "The list of spaces accessible by the authenticated user"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "List spaces",
        "tags": [
          "Spaces"
        ]
      },
      "post": {
        "description": "Creates a new space based on a supplied name and description in the JSON request body. The endpoint is subject to rate limiting and automatically applies default configurations such as generating a random icon and initializing the workspace with default settings (for example, a default dashboard or home page). On success, the new space’s full metadata is returned, enabling the client to immediately switch context to the new internal.",
        "operationId": "create_space",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpaceRequest"
              }
            }
          },
          "description": "The space to create",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            },
            "description": "The created space"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "Create space",
        "tags": [
          "Spaces"
        ]
      }
    },
    "/v1/spaces/{space_id}": {
      "get": {
        "description": "Fetches full details about a single space identified by its space ID. The response includes metadata such as the space name, icon, and various workspace IDs (home, archive, profile, etc.). This detailed view supports use cases such as displaying space-specific settings.",
        "operationId": "get_space",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          },
          {
            "description": "The ID of the space to retrieve; must be retrieved from ListSpaces endpoint",
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            },
            "description": "The space details"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            },
            "description": "Space not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "Get space",
        "tags": [
          "Spaces"
        ]
      },
      "patch": {
        "description": "Updates the name or description of an existing space. The request body should contain the new name and/or description in JSON format. This endpoint is useful for renaming or rebranding a workspace without needing to recreate it. The updated space’s metadata is returned in the response.",
        "operationId": "update_space",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          },
          {
            "description": "The ID of the space to update; must be retrieved from ListSpaces endpoint",
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpaceRequest"
              }
            }
          },
          "description": "The space details to update",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            },
            "description": "The updated space"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            },
            "description": "Space not found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "Update space",
        "tags": [
          "Spaces"
        ]
      }
    },
...
    "/v1/spaces/{space_id}/lists/{list_id}/views": {
      "get": {
        "description": "Returns a paginated list of views defined for a specific list (query or collection) within a space. Each view includes details such as layout, applied filters, and sorting options, enabling clients to render the list according to user preferences and context. This endpoint is essential for applications that need to display lists in various formats (e.g., grid, table) or with different sorting/filtering criteria.",
        "operationId": "get_list_views",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          },
          {
            "description": "The ID of the space to which the list belongs; must be retrieved from ListSpaces endpoint",
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The ID of the list to retrieve views for; must be retrieved from SearchSpace endpoint with types: ['collection', 'set']",
            "in": "path",
            "name": "list_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The number of items to skip before starting to collect the result set",
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "type": "integer"
            }
          },
          {
            "description": "The number of items to return",
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse-View"
                }
              }
            },
            "description": "The list of views associated with the specified list"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            },
            "description": "Not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "Get list views",
        "tags": [
          "Lists"
        ]
      }
    },
...
  },
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "http://127.0.0.1:31009"
    }
  ]
}